% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGWP.R
\name{getGWP}
\alias{getGWP}
\title{Retrieves GWP data from .csv file}
\usage{
getGWP(source = "AR6", metric = "GWP100", species = NULL)
}
\arguments{
\item{source}{name of the gwp data source, choose from "AR6" (default), "SAR", "AR4", "AR5",}

\item{metric}{gwp metric, indicating the years after emission that are being considered, choose from
"GWP100" (default), "GWP20", "GWP500", "CCFGWP100", GTP100"}
}
\description{
Retrieves GWP data from .csv file
}
\author{
Pascal Weigmann
}
